requirements.txt
anyio==3.7.0
APScheduler==3.6.3
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
Pillow==9.5.0
python-telegram-bot==13.14
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tornado==6.1
typing_extensions==4.6.3
tzlocal==5.0.1
urllib3==2.0.3
